Split up the card & column actions
